Move kernel/test/_backend/_helpers into geometry/ subpackage
Make pyriemann.geometry a fully standalone subpackage by relocating the
remaining cross-package dependencies into it:

- pyriemann.utils.kernel       -> pyriemann.geometry.kernel
- pyriemann.utils.test         -> pyriemann.geometry.test
- pyriemann.utils._backend     -> pyriemann.geometry._backend (private)
- pyriemann._helpers           -> pyriemann.geometry._helpers (private)

After this change, modules under pyriemann/geometry/ only import from
each other (sibling-relative `from ._backend import ...`,
`from ._helpers import ...`).

Backward compatibility:
- Public modules (utils.kernel, utils.test) retain shim files that
  re-export from the new location and emit a DeprecationWarning.
- Private modules (utils._backend, _helpers) get silent re-export shims
  (they were never part of the public API).

Internal callers (classification, regression, embedding, datasets/sampling,
utils/_check, utils/docs) updated to import from the new geometry paths so
the deprecation warnings only fire for end users.

Test files for moved modules renamed test_utils_* -> test_geometry_*.

doc/api.rst: split top-level "Utility Functions" section into "Geometry
Functions" (covariance/distance/mean/median/geodesic/kernel/tangent space/
base/ajd/matrix tests) and "Utility Functions" (visualization), update
currentmodule directives for moved modules.

doc/whatsnew.rst: extend v0.12.dev entry to cover the additional moves.
