+++ pyriemann/embedding.py
@@ -10,7 +10,7 @@ from sklearn.manifold import spectral_embedding
 from sklearn.manifold._utils import _binary_search_perplexity
 
 from .geometry.distance import pairwise_distance
-from .utils.kernel import kernel as kernel_fct
+from .geometry.kernel import kernel as kernel_fct
 from .optimization.positive_definite import _get_tsne_embedding
 
 
@@ -151,7 +151,7 @@ class LocallyLinearEmbedding(TransformerMixin, BaseEstimator):
         n_matrices - 1.
     metric : string, default="riemann"
         Metric used for k-NN and kernel estimation. For the list of supported
-        metrics, see :func:`pyriemann.utils.kernel.kernel`.
+        metrics, see :func:`pyriemann.geometry.kernel.kernel`.
     kernel : callable | None, default=None
         Kernel function to use for the embedding. If None, the canonical
         kernel specified by the metric is used. Must be a function that
@@ -414,7 +414,7 @@ def barycenter_weights(X, Y, indices, metric="riemann", kernel=None, reg=1e-3):
         Indices of matrices in Y used to compute the barycenter.
     metric : string, default="riemann"
         Kernel metric. For the list of supported metrics, see
-        :func:`pyriemann.utils.kernel.kernel`.
+        :func:`pyriemann.geometry.kernel.kernel`.
     kernel : callable | None, default=None
         Kernel function to use for the embedding. If None, the canonical
         kernel specified by the metric is used. Must be a function that
@@ -492,7 +492,7 @@ def locally_linear_embedding(
         ``n_matrices - 1``.
     metric : string, default="riemann"
         Metric used for k-NN and kernel estimation. For the list of supported
-        metrics, see :func:`pyriemann.utils.kernel.kernel`.
+        metrics, see :func:`pyriemann.geometry.kernel.kernel`.
     kernel : callable | None, default=None
         Kernel function to use for the embedding. If None, the canonical
         kernel specified by the metric is used. Must be a function that
