+++ pyriemann/geometry/covariance.py
@@ -5,21 +5,21 @@ from array_api_compat import array_namespace as get_namespace, device as xpd
 from array_api_extra import expand_dims
 from scipy.stats import chi2
 
-from pyriemann._helpers import (
-    check_function,
-    check_init,
-    check_weights,
-    deprecated,
-    is_real_type,
-    is_square,
-)
-from pyriemann.utils._backend import (
+from ._backend import (
     apply_xp_cov,
     diag_indices,
     hann_window,
     fast_mcd,
     ledoit_wolf,
     oas,
+)
+from ._helpers import (
+    check_function,
+    check_init,
+    check_weights,
+    deprecated,
+    is_real_type,
+    is_square,
 )
 from .base import ctranspose, _vectorize_nd
 from .distance import distance_mahalanobis
