+++ pyriemann/geometry/geodesic.py
@@ -2,8 +2,8 @@
 
 from array_api_compat import array_namespace as get_namespace
 
-from pyriemann._helpers import check_function, check_matrix_pair
-from pyriemann.utils._backend import diag_indices, tril_indices
+from ._backend import diag_indices, tril_indices
+from ._helpers import check_function, check_matrix_pair
 from .base import ctranspose, _eigvalsh, sqrtm, invsqrtm, powm, logm, expm
