+++ pyriemann/utils/docs.py
@@ -1,5 +1,5 @@
 """Backward-compatibility re-export.
 
-The ``deprecated`` decorator now lives in :mod:`pyriemann._helpers`.
+The ``deprecated`` decorator now lives in :mod:`pyriemann.geometry._helpers`.
 """
-from pyriemann._helpers import deprecated  # noqa: F401
+from ..geometry._helpers import deprecated  # noqa: F401
