+++ tests/test_classification.py
@@ -11,7 +11,7 @@ from sklearn.model_selection import cross_val_score, KFold
 from sklearn.pipeline import make_pipeline
 
 from pyriemann.estimation import Covariances
-from pyriemann.utils.kernel import kernel
+from pyriemann.geometry.kernel import kernel
 from pyriemann.geometry.mean import gmean
 from pyriemann.classification import (
     _mode_2d,
