+++ tests/test_estimation.py
@@ -19,7 +19,7 @@ from pyriemann.estimation import (
     BlockCovariances,
     Kernels,
 )
-from pyriemann.utils.test import (
+from pyriemann.geometry.test import (
     is_sym_pos_def as is_spd,
     is_sym_pos_semi_def as is_spsd,
     is_herm_pos_def as is_hpd,
