+++ tests/test_geometry_base.py
@@ -22,7 +22,7 @@ from pyriemann.geometry.base import (
     ddexpm,
     ddlogm,
 )
-from pyriemann.utils.test import is_pos_def, is_sym_pos_def, is_hermitian
+from pyriemann.geometry.test import is_pos_def, is_sym_pos_def, is_hermitian
 
 
 n_channels = 3
