+++ tests/test_geometry_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from pyriemann.utils.test import (
+from pyriemann.geometry.test import (
     is_square, is_sym, is_skew_sym, is_hankel,
     is_real, is_real_type, is_hermitian,
     is_pos_def, is_pos_semi_def,
