+++ tests/test_sampling.py
@@ -4,7 +4,7 @@ import pytest
 
 from pyriemann.datasets import sample_gaussian_spd, RandomOverSampler
 from pyriemann.geometry.distance import distance_riemann
-from pyriemann.utils.test import is_sym_pos_def as is_spd
+from pyriemann.geometry.test import is_sym_pos_def as is_spd
 
 
 @pytest.mark.parametrize("n_jobs", [1, -1])
